==> 05_Interfacing_network_analysis_with_other_data_such_as_functional_annotation_and_gene_ontology.Rmd <==
# Interfacing network analysis with other data such as functional annotation and gene ontology

```{r Code chunk 1 5, include=FALSE}
# Display the current working directory
getwd();
# If necessary, change the path below to the directory where the data files are stored. 
# "." means current directory. On Windows use a forward slash / instead of the usual \.
workingDir = ".";
setwd(workingDir); 
# Load the WGCNA package
library(WGCNA)
# The following setting is important, do not omit.
options(stringsAsFactors = FALSE);
# Load the data saved in the first part
lnames = load(file = "saved_data/name_of_analysis.RData");
lnames
lnames = load(file = paste("saved_data/01",name_of_analysis,"_data_input.RData", sep=""));
# Load network data saved in the second part.
lnames = load(file = paste("saved_data/02",name_of_analysis,"_networkConstruction_auto.RData", sep=""));
lnames
lnames = load(file = paste("./saved_data/04",name_of_analysis,"_relating_modules_to_outcome2.RData", sep=""));
lnames
```

```{r Code chunk 1.5 5, include=FALSE}
library(org.Hs.eg.db)
hs <- org.Hs.eg.db
my.symbols = names(dat_expr)
length(my.symbols)
ei <- select(hs, 
             keys = my.symbols,
             columns = c("ENTREZID", "SYMBOL"),
             keytype = "SYMBOL")
sum(is.na(ei$ENTREZID)) #count missing values in col
dim(ei)
ei0.0 <- na.omit(ei)
dim(ei0.0)
ei[1:5,]

```

 
## For each interesting module, create a file with its genes 

A file with all genes as background for enrichment analysis is created and files for genes in interesting modules.


```{r Code chunk 2 5, include=FALSE}
my.symbols = names(dat_expr)

symb_orig2symb_biom_out = match(my.symbols, ei$SYMBOL)
# The following is the number or probes without annotation:
sum(is.na(symb_orig2symb_biom_out))
# Should return 0.

# Get the corresponding Locuis Link IDs
allLLIDs = ei$ENTREZID[symb_orig2symb_biom_out];
# $ Choose interesting modules
intModules = module_col_of_int
for (module in intModules)
{
  # Select module probes
  modGenes = (moduleColors==module)
  # Get their entrez ID codes
  modLLIDs = allLLIDs[modGenes];
  # Write them into a file
  fileName = paste("./out",name_of_analysis,"/LocusLinkIDs-", module, ".txt", sep="");
  write.table(as.data.frame(modLLIDs), file = fileName,
              row.names = FALSE, col.names = FALSE)
}
# As background in the enrichment analysis, we will use all probes in the analysis.
fileName = paste("./out",name_of_analysis,"/LocusLinkIDs-all.txt", sep="");
write.table(as.data.frame(allLLIDs), file = fileName,
            row.names = FALSE, col.names = FALSE)
```


## Performing Gene Ontology enrichment for each module

```{r Code chunk 3.5 5, include=FALSE}
#library(anRichment)
if (file.exists(paste("saved_data/04",name_of_analysis,"_GOenr.RData", sep=""))){
        lnames = load(file = paste("saved_data/04",name_of_analysis,"_GOenr.RData", sep=""));
        lnames  
} else{
GOenr = GOenrichmentAnalysis(moduleColors, allLLIDs, organism = "human", nBestP = 10);
save(GOenr, file = paste("saved_data/04",name_of_analysis,"_GOenr.RData", sep=""))
}
#GOenr = enrichmentAnalysis(moduleColors, allLLIDs, refCollection=buildGOcollection(organism = "mouse"));

# http://bioconductor.org/packages/release/data/annotation/html/org.Mm.eg.db.htmls
#install mouse genome
#if (!requireNamespace("BiocManager", quietly = TRUE))
#  install.packages("BiocManager")
#BiocManager::install("org.Mm.eg.db")
# Gives error:
# ERROR: compilation failed for package ‘WGCNA’
```

```{r Code chunk 4 5, include=FALSE}
tab = GOenr$bestPTerms[[4]]$enrichment
```

```{r Code chunk 5 5}
names(tab)

```

## Write GOenrichment table to file

```{r Code chunk 6 5, include=FALSE}
fileName = paste("./out",name_of_analysis,"/GOEnrichmentTable.csv", sep="");
write.table(tab, file = fileName, sep = ",", quote = TRUE, row.names = FALSE)

```

```{r Code chunk 7 5, include=FALSE}
keepCols = c(1, 2, 5, 6, 7, 12, 13);
screenTab = tab[, keepCols];
# Round the numeric columns to 2 decimal places:
numCols = c(3, 4);
screenTab[, numCols] = signif(apply(screenTab[, numCols], 2, as.numeric), 2)
# Truncate the the term name to at most 40 characters
screenTab[, 7] = substring(screenTab[, 7], 1, 40)
# Shorten the column names:
colnames(screenTab) = c("module", "size", "p-val", "Bonf", "nInTerm", "ont", "term name");
rownames(screenTab) = NULL;
# Set the width of R's output. The reader should play with this number to obtain satisfactory output.
options(width=95)
# Finally, display the enrichment table:
screenTab
```
